# Conflicts:
#	src/client/components/SettingUI.js
#	src/server/routes/memberJoin.js

+++ src/server/routes/scrap.js
@@ -93,6 +93,46 @@ scrap.post('/', (req,res)=>{
                     return res.send('error' + err)
                 })
         }else{
+            return res.status(400).json({ // HTTP 요청에 대한 리스폰스 (json 형식으로)
+                error: "duplicate scrap",
+                code: 1
+            });
+            /*MemberScrap.destroy({
+                where:{
+                    user_id : scrapData.user_id,
+                    recipe_code : scrapData.recipe_code
+                }
+            })
+                .then(memberScrap=>{
+                    console.log("scrap delete");
+                    return res.json({success: true})
+                })
+                .catch(err=>{
+                    console.log("scrap error");
+                    return res.send('error' + err)
+                })*/
+        }
+    }).catch((err)=>{
+        return res.send('error' + err);
+    })
+});
+
+scrap.post('/delete', (req,res)=>{
+    console.log("scrap delete post route");
+
+    const scrapData = {
+        user_id : req.body.user_id,
+        recipe_code : req.body.recipe_code,
+        SCRAP_DATE : req.body.SCRAP_DATE
+    };
+    console.log("scrap post route2");
+    MemberScrap.findOne({
+        where :{
+            user_id : scrapData.user_id,
+            recipe_code : scrapData.recipe_code
+        }
+    }).then((memberScrap)=>{
+        if(memberScrap){
             MemberScrap.destroy({
                 where:{
                     user_id : scrapData.user_id,
